requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.2
openai>=1.24.0,<2.0.0
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
faiss-cpu==1.9.0.post1
python-dateutil==2.9.0

